+++ source/Squery/BuilderProxy.php
@@ -2,6 +2,7 @@
 
 namespace League\Squery;
 
+use League\Squery;
 use League\Squery\Builder\Builder;
 use League\Squery\Factory\BuilderFactory;
 
@@ -21,6 +22,8 @@ class BuilderProxy extends AbstractProxy
     {
         if ($factory === null) {
             $factory = new BuilderFactory();
+        } else {
+            Squery::depreciate("Factories are depreciated (http://squery.thephpleague.com/services)");
         }
 
         $this->factory = $factory;