+++ source/Squery/Factory/ProcessFactory.php
@@ -2,9 +2,9 @@
 
 namespace League\Squery\Factory;
 
+use League\Squery;
 use League\Squery\Factory;
-use League\Squery\Process\SymfonyProcess;
-use Symfony\Component\Process\Process;
+use League\Squery\Process\Process;
 
 class ProcessFactory implements Factory
 {
@@ -13,8 +13,6 @@ class ProcessFactory implements Factory
      */
     public function newInstance()
     {
-        return new SymfonyProcess(
-            new Process("echo")
-        );
+        return Squery::container()["squery/process"];
     }
 }