+++ tests/Squery/RunnerProxyTest.php
@@ -1,12 +1,14 @@
 <?php
 
-namespace League\Squery\Tests;
+namespace League\Tests\Squery;
 
 use League\Squery\Builder\Builder;
-use League\Squery\BuilderProxy;
-use League\Squery\Factory\BuilderFactory;
+use League\Squery\Factory\RunnerFactory;
+use League\Squery\Runner\Runner;
+use League\Squery\RunnerProxy;
+use Mockery;
 
-class BuilderProxyTest extends TestCase
+class RunnerProxyTest extends TestCase
 {
     use Traits\Proxy\CustomFactories;
     use Traits\Proxy\DefaultFactories;
@@ -22,7 +24,7 @@ class BuilderProxyTest extends TestCase
      */
     public function itCreatesDefaultFactory()
     {
-        $this->assertCustomFactories(BuilderFactory::class, BuilderProxy::class);
+        $this->assertCustomFactories(RunnerFactory::class, RunnerProxy::class);
     }
 
     /**
@@ -32,7 +34,7 @@ class BuilderProxyTest extends TestCase
      */
     public function itAllowsCustomFactories()
     {
-        $this->assertCustomFactories(BuilderFactory::class, BuilderProxy::class);
+        $this->assertCustomFactories(RunnerFactory::class, RunnerProxy::class);
     }
 
     /**
@@ -42,7 +44,7 @@ class BuilderProxyTest extends TestCase
      */
     public function itThrowsForMissingStaticMethods()
     {
-        $this->assertMissingStaticMethods(BuilderProxy::class);
+        $this->assertMissingStaticMethods(RunnerProxy::class);
     }
 
     /**
@@ -52,7 +54,7 @@ class BuilderProxyTest extends TestCase
      */
     public function itThrowsForMissingMethods()
     {
-        $this->assertMissingMethods(BuilderProxy::class);
+        $this->assertMissingMethods(RunnerProxy::class);
     }
 
     /**
@@ -62,7 +64,7 @@ class BuilderProxyTest extends TestCase
      */
     public function itThrowsForInvalidFactories()
     {
-        $this->assertInvalidFactories(BuilderProxy::class, "select");
+        $this->assertInvalidFactories(RunnerProxy::class, "select");
     }
 
     /**
@@ -72,6 +74,8 @@ class BuilderProxyTest extends TestCase
      */
     public function itDelegatesMethods()
     {
-        $this->assertDelegatesMethods(BuilderFactory::class, Builder::class, "select", BuilderProxy::class, "select");
+        $builder = Mockery::mock(Builder::class);
+
+        $this->assertDelegatesMethods(RunnerFactory::class, Runner::class, "run", RunnerProxy::class, "run", [$builder]);
     }
 }